Casted size_t to float

+++ Libs/Engine/sources/Scenes/Scenes.cpp
@@ -95,7 +95,7 @@ void Scenes::InitMap(ECS::Coordinator& coordinator, RayLib::Camera3D& camera, co
             }
         }
     }
-    ECS::Entity& mapFloor = entityFactory.createFloor(RayLib::Vector2<float>(map[0].size(), map.size()));
+    ECS::Entity& mapFloor = entityFactory.createFloor(RayLib::Vector2<float>(static_cast<float>(map[0].size()), static_cast<float>(map.size())));
     mapFloor.SetTag("MapFloor");
 }
